Switch VLM backends from bf16 to fp16 for RDNA3 safety
bf16 causes GPU hangs (HW Exception) on ROCm/RDNA3 with certain model
architectures (confirmed on Qwen3-ASR audio encoder Conv2d layers).
fp16 has native 2x throughput on RDNA3 and works on all tested models.

+++ vsg_core/subtitles/ocr/vlm_backends/lfm2vl.py
@@ -45,7 +45,7 @@ class LFM2VLBase(VLMBackend):
         self.processor = AutoProcessor.from_pretrained(model_path)
         self.model = AutoModelForImageTextToText.from_pretrained(
             model_path,
-            dtype=torch.bfloat16,
+            dtype=torch.float16,
             low_cpu_mem_usage=True,
             attn_implementation="sdpa",
             device_map="cpu",
